Basic Functionality
I have added the basci functionality to our dog website. This includes
the html with a title, button, and a div for storing the results. The
javascript fetches the api and creates the appropiate element to show
the image/video. The css just resizes images and videos to 50% width and
auto height.